refactor demo code
