+++ demo/2.link-self.php
@@ -1,0 +1,72 @@
+<?php
+/**
+ * This file is part of the BEAR.Resource package.
+ *
+ * @license http://opensource.org/licenses/MIT MIT
+ */
+namespace Sandbox\Resource\Resource\App;
+
+use BEAR\Resource\Annotation\Link;
+use BEAR\Resource\Module\HalModule;
+use BEAR\Resource\Module\ResourceModule;
+use BEAR\Resource\ResourceInterface;
+use BEAR\Resource\ResourceObject;
+use Ray\Di\Injector;
+
+require dirname(__DIR__) . '/vendor/autoload.php';
+
+class Blog extends ResourceObject
+{
+    protected $users = [
+        ['name' => 'Athos blog'],
+        ['name' => 'Aramis blog'],
+        ['name' => 'Porthos blog']
+    ];
+
+    public function onGet($id)
+    {
+        $this->body = $this->users[$id];
+
+        return $this;
+    }
+}
+
+class User extends ResourceObject
+{
+    private $users = [
+        ['name' => 'Athos', 'age' => 15, 'blog_id' => 0],
+        ['name' => 'Aramis', 'age' => 16, 'blog_id' => 1],
+        ['name' => 'Porthos', 'age' => 17, 'blog_id' => 2]
+    ];
+
+    /**
+     * @Link(rel="blog", href="app://self/blog?id={blog_id}")
+     */
+    public function onGet($id)
+    {
+        return $this->users[$id];
+    }
+}
+
+/* @var $resource \BEAR\Resource\ResourceInterface */
+$resource = (new Injector(new HalModule(new ResourceModule('Sandbox\Resource'))))->getInstance(ResourceInterface::class);
+$user = $resource->get->uri('app://self/user')(['id' => 2]);
+
+echo $user;
+
+//{
+//    "name": "Porthos",
+//    "age": 17,
+//    "blog_id": 2,
+//    "blog": {
+//    "name": "Porthos blog"
+//    },
+//    "_links": {
+//    "self": {
+//        "href": "/user?id=2"
+//        },
+//        "blog": {
+//        "href": "app://self/blog?id=2"
+//        }
+//    }
+//}
