+++ demo/run.php
@@ -1,0 +1,12 @@
+<?php
+/**
+ * This file is part of the BEAR.Resource package.
+ *
+ * @license http://opensource.org/licenses/MIT MIT
+ */
+passthru('php ' . __DIR__ . '/0.min.php');
+passthru('php ' . __DIR__ . '/1.basic.php');
+passthru('php ' . __DIR__ . '/2.link-self.php');
+passthru('php ' . __DIR__ . '/3.link-crawl.php');
+passthru('php ' . __DIR__ . '/4.restbucks.php');
+passthru('php ' . __DIR__ . '/5.embed.php');
