--- docs/demo/02.link-self/Sandbox/Resource/App/Blog.php
@@ -1,25 +1,0 @@
-<?php
-/**
- * This file is part of the BEAR.Resource package.
- *
- * @license http://opensource.org/licenses/MIT MIT
- */
-namespace Sandbox\Resource\Resource\App;
-
-use BEAR\Resource\ResourceObject;
-
-class Blog extends ResourceObject
-{
-    protected $users = [
-        ['name' => 'Athos blog'],
-        ['name' => 'Aramis blog'],
-        ['name' => 'Porthos blog']
-    ];
-
-    public function onGet($id)
-    {
-        $this->body = $this->users[$id];
-
-        return $this;
-    }
-}
