--- docs/demo/04.restbucks/scripts/instance.php
@@ -1,18 +1,0 @@
-<?php
-/**
- * This file is part of the BEAR.Resource package.
- *
- * @license http://opensource.org/licenses/MIT MIT
- */
-use BEAR\Resource\Module\ResourceModule;
-use BEAR\Resource\ResourceInterface;
-use Ray\Di\Injector;
-
-$loader = require dirname(dirname(dirname(dirname(__DIR__)))) . '/vendor/autoload.php';
-$loader->add('', dirname(__DIR__));
-
-$injector = new Injector(new ResourceModule('Restbucks'));
-$resource = $injector->getInstance(ResourceInterface::class);
-
-/* @var $resource \BEAR\Resource\ResourceInterface */
-return $resource;
