--- docs/demo/04.restbucks/src.php
@@ -1,9 +1,0 @@
-<?php
-/**
- * This file is part of the BEAR.Resource package.
- *
- * @license http://opensource.org/licenses/MIT MIT
- */
-$loader = require 'vendor/autoload.php';
-/* @var $loader \Composer\Autoload\ClassLoader */
-$loader->add('Restbucks', [dirname(__DIR__)]);
