--- docs/demo/10.cache/Resource/App/Weather.php
@@ -1,23 +1,0 @@
-<?php
-/**
- * This file is part of the BEAR.Resource package.
- *
- * @license http://opensource.org/licenses/MIT MIT
- */
-namespace MyVendor\MyApp\Resource\App;
-
-use BEAR\Resource\ResourceObject;
-
-class Weather extends ResourceObject
-{
-    public $links = [
-        'tomorrow' => 'app://self/weather/tomorrow'
-    ];
-
-    public function onGet($date)
-    {
-        $this['today'] = "the weather of {$date} is sunny";
-
-        return $this;
-    }
-}
